requirements.txt
Django
psycopg[binary]
python-dotenv
openai
celery
redis